+++ lib/src/models/side_menu_item_with_global_base.dart
@@ -1,0 +1,5 @@
+import 'package:flutter/widgets.dart';
+
+abstract class SideMenuItemWithGlobalBase extends StatefulWidget {
+  const SideMenuItemWithGlobalBase({Key? key}) : super(key: key);
+}