+++ lib/src/side_menu_expansion_item.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:easy_sidemenu/src/side_menu_item.dart';
 import 'package:easy_sidemenu/src/side_menu_controller.dart';
+import 'package:easy_sidemenu/src/models/side_menu_item_type.dart';
 
-class SideMenuExpansionItem {
+class SideMenuExpansionItem implements SideMenuItemType {
   /// #### Side Menu Item
   ///
   /// This is a widget as [SideMenu] items with text and icon