+++ lib/src/side_menu_item.dart
@@ -1,8 +1,9 @@
+import 'package:easy_sidemenu/src/models/side_menu_item_type.dart';
 import 'package:flutter/material.dart';
 import 'package:easy_sidemenu/src/side_menu_display_mode.dart';
 import 'package:easy_sidemenu/src/side_menu_controller.dart';
 
-class SideMenuItem {
+class SideMenuItem implements SideMenuItemType {
   /// Title text
   final String? title;